vsx: support buildtime 'extensionPack' handling
The following commit adds improved support for `extensionPack` plugins defined as
builtins. Previously, `extensionPacks` could be referenced as builtins,
but they would later be resolved when the application is first started.
For large packs (such as vscode builtins) this would result in a long
startup time, where extensionPacks would need to be resolved, and then
downloaded. The following commit fixes that problem by allowing
`extensionPack` to be resolved and fetched at buildtime rather than
runtime.

In order to implement the improvement, a new dev-package was created
(`@theia/ovsx-client`) which includes the common logic for the `cli` and
`@theia/vsx-registry` (previously part of the `@theia/vsx-registry`)
extension. This logic which interacts with the registry (fetch, search,
download, determine compatibility) is now used by the `cli` (buildtime),
and `vsx-registry` (runtime). The `@theia/vsx-registry` extension was
refactored and cleaned up as a consequence.

The idea when resolving builtin extensions is the following:
- download extensions like today (master).
- determine if any `extensionPack`(s) exist.
- fetch individual extension ids.
- download latest compatible versions of all individual extension ids.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>
Co-authored-by: marechal-p <paul.marechal@ericsson.com>